Extract a shared `ButtonEdit` component.
