+++ blocks/reader-revenue-manager/subscribe-with-google/Edit.js
@@ -17,17 +17,14 @@
 /**
  * WordPress dependencies
  */
-import { useBlockProps, InspectorControls } from '@wordpress-core/block-editor';
-import { ExternalLink, Notice } from '@wordpress-core/components';
-import { createInterpolateElement, Fragment } from '@wordpress-core/element';
+import { ExternalLink } from '@wordpress-core/components';
+import { createInterpolateElement } from '@wordpress-core/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { EditorButton } from '../common';
-import { CORE_EDITOR } from '../common/constants';
-import { CORE_MODULES } from '../../../assets/js/googlesitekit/modules/datastore/constants';
+import { ButtonEdit } from '../common';
 import { MODULES_READER_REVENUE_MANAGER } from '../../../assets/js/modules/reader-revenue-manager/datastore/constants';
 
 /**
@@ -40,15 +37,9 @@ import { MODULES_READER_REVENUE_MANAGER } from '../../../assets/js/modules/reade
  * @return {Element} Element to render.
  */
 export default function Edit( { select } ) {
-	const blockProps = useBlockProps();
-
-	const hasModuleAccess = select( CORE_MODULES ).hasModuleOwnershipOrAccess(
-		'reader-revenue-manager'
-	);
-
-	const settings = select( MODULES_READER_REVENUE_MANAGER ).getSettings();
-
-	const { publicationID, paymentOption, snippetMode, postTypes } = settings;
+	const publicationID = select(
+		MODULES_READER_REVENUE_MANAGER
+	).getPublicationID();
 
 	const serviceURL = select( MODULES_READER_REVENUE_MANAGER ).getServiceURL( {
 		path: 'reader-revenue-manager',
@@ -57,21 +48,14 @@ export default function Edit( { select } ) {
 		},
 	} );
 
-	const metaKey = `googlesitekit_rrm_${ publicationID }:productID`;
-
-	const postProductID =
-		select( CORE_EDITOR ).getEditedPostAttribute( 'meta' )?.[ metaKey ] ||
-		'';
-	const postType = select( CORE_EDITOR ).getCurrentPostType();
-
-	let notice = '';
-	let disabled = false;
-
-	if ( paymentOption !== 'subscriptions' ) {
-		disabled = true;
-
-		if ( hasModuleAccess ) {
-			notice = createInterpolateElement(
+	return (
+		<ButtonEdit
+			buttonLabel={
+				/* translators: Button label for Subscribe with Google. See: https://github.com/subscriptions-project/swg-js/blob/05af2d45cfcaf831a6b4d35c28f2c7b5c2e39308/src/i18n/swg-strings.ts#L24-L57 (please refer to the latest version of the file) */
+				__( 'Subscribe with Google', 'google-site-kit' )
+			}
+			requiredPaymentOption="subscriptions"
+			invalidPaymentOptionWithModuleAccessNotice={ createInterpolateElement(
 				__(
 					'You need to set up a paywall in Reader Revenue Manager to use this block. <a>Go to Reader Revenue Manager</a>',
 					'google-site-kit'
@@ -79,24 +63,12 @@ export default function Edit( { select } ) {
 				{
 					a: <ExternalLink href={ serviceURL } />,
 				}
-			);
-		} else {
-			notice = __(
+			) }
+			invalidPaymentOptionWithoutModuleAccessNotice={ __(
 				'You need to set up a paywall in Reader Revenue Manager to use this block. Contact your administrator.',
 				'google-site-kit'
-			);
-		}
-	} else if (
-		postProductID === 'none' ||
-		( ! postProductID && snippetMode === 'per_post' ) ||
-		( ! postProductID &&
-			snippetMode === 'post_types' &&
-			! postTypes.includes( postType ) )
-	) {
-		disabled = true;
-
-		if ( hasModuleAccess ) {
-			notice = createInterpolateElement(
+			) }
+			noSnippetWithModuleAccessNotice={ createInterpolateElement(
 				__(
 					'This post does not include the Reader Revenue Manager snippet. Configure the snippet for this post in the post settings sidebar.',
 					'google-site-kit'
@@ -104,36 +76,11 @@ export default function Edit( { select } ) {
 				{
 					a: <ExternalLink href={ serviceURL } />,
 				}
-			);
-		} else {
-			notice = __(
+			) }
+			noSnippetWithoutModuleAccessNotice={ __(
 				'This post does not include the Reader Revenue Manager snippet. Contact your administrator',
 				'google-site-kit'
-			);
-		}
-	}
-
-	return (
-		<Fragment>
-			{ notice && (
-				<InspectorControls>
-					<div className="block-editor-block-card">
-						<Notice status="warning" isDismissible={ false }>
-							{ notice }
-						</Notice>
-					</div>
-				</InspectorControls>
 			) }
-			<div { ...blockProps }>
-				<div className="googlesitekit-blocks-reader-revenue-manager">
-					<EditorButton disabled={ disabled }>
-						{
-							/* translators: Button label for Subscribe with Google. See: https://github.com/subscriptions-project/swg-js/blob/05af2d45cfcaf831a6b4d35c28f2c7b5c2e39308/src/i18n/swg-strings.ts#L24-L57 (please refer to the latest version of the file) */
-							__( 'Subscribe with Google', 'google-site-kit' )
-						}
-					</EditorButton>
-				</div>
-			</div>
-		</Fragment>
+		/>
 	);
 }
